+++ lib/features/bonfire/view_models/bonfire_state.dart
@@ -1,0 +1,12 @@
+
+
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'bonfire_state.freezed.dart';
+
+@freezed
+class BonfireState with _$BonfireState {
+  const factory BonfireState({
+     int? selectedOptionIndex,
+  }) = _BonfireState;
+}